Add active creators community data model
This commit is a data model (series of SQL statements) that can be used to build an active creators table. This active creators table is one row per user per day and summarises a wide range of fields that would otherwise be hard to consolidate in a single manually written query.